Revert "add remote DataStore"
