+++ README.md
@@ -2,7 +2,7 @@
 
 # Getting started.
 
-Build and run the project:
+Build and run the project: 
 ```
 git clone --recurse-submodules https://github.com/WingsDao/move-vm-in-cosmos.git
 cd move-vm-in-cosmos
@@ -12,11 +12,9 @@ Run:
 
 Format: `client` or `server` `IP:PORT`
 
-DS Mock Server: `cargo run --bin ds-server "[::1]:50052"`
+Server: `cargo run --bin server "[::1]:50051"`
 
-VM Server: `cargo run --bin server "[::1]:50051" "http://[::1]:50052"`
-
-VM Mock Client: `cargo run --bin client "http://[::1]:50051"`
+Client: `cargo run --bin client "http://[::1]:50051"`
 
 # precommit hook
